Remove unnecessary CRC hash
Also move MD5 to android only
