+++ Xamarin.Essentials/Types/Shared/Utils.shared.cs
@@ -1,7 +1,4 @@
 ﻿using System;
-#if !NETSTANDARD1_0
-using System.Security.Cryptography;
-#endif
 using System.Text;
 using System.Threading;
 using System.Threading.Tasks;
@@ -19,22 +16,6 @@ namespace Xamarin.Essentials
                 return new Version(major, 0);
 
             return new Version(0, 0);
-        }
-
-        internal static string Md5Hash(string input)
-        {
-#if NETSTANDARD1_0
-            throw new NotImplementedInReferenceAssemblyException();
-#else
-            var hash = new StringBuilder();
-            var md5provider = new MD5CryptoServiceProvider();
-            var bytes = md5provider.ComputeHash(Encoding.UTF8.GetBytes(input));
-
-            for (var i = 0; i < bytes.Length; i++)
-                hash.Append(bytes[i].ToString("x2"));
-
-            return hash.ToString();
-#endif
         }
 
         internal static CancellationToken TimeoutToken(CancellationToken cancellationToken, TimeSpan timeout)
